+++ webui/src/change-password.vue
@@ -38,6 +38,7 @@
               v-model="currentPassword"
               :placeholder="t('changePassword.currentPasswordPlaceholder')"
               :state="v$.currentPassword.$error ? false : null"
+              autocomplete="current-password"
               size="lg"
             />
             <BFormInvalidFeedback v-if="v$.currentPassword.required.$invalid">
@@ -51,6 +52,7 @@
               v-model="newPassword"
               :placeholder="t('changePassword.newPasswordPlaceholder')"
               :state="v$.newPassword.$error ? false : null"
+              autocomplete="new-password"
               size="lg"
             />
             <BFormInvalidFeedback v-if="v$.newPassword.minLength.$invalid">
@@ -67,6 +69,7 @@
               v-model="confirmPassword"
               :placeholder="t('changePassword.confirmPasswordPlaceholder')"
               :state="v$.confirmPassword.$error ? false : null"
+              autocomplete="new-password"
               size="lg"
               @keyup.enter="handleSubmit"
             />