+++ webui/src/settings.vue
@@ -58,6 +58,7 @@
                     type="text"
                     autocomplete="username"
                     autocapitalize="none"
+                    autocorrect="off"
                     spellcheck="false"
                     :state="v$.adminUsername.$error ? false : null"
                   />